Web/requirements.txt
Flask
psycopg2-binary
requests
beautifulsoup4
